+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.32.2
-- **[fix]** : `NavigationMenu`  - Active link mobile
+## v1.33.0
+- **[feat]** : `NavigationMenu`  - Add loading state for `secondaryMenu`
+- **[refactor]** : `NavigationMenu`  - Improve contrast for `secondaryMenu`
+- **[refactor]** : `NavigationMenu`  - Change background color avatar